update relation et err msg en suppression

+++ src/main/java/com/example/appgestiondeprojet/services/ProjetServiceImpl.java
@@ -5,9 +5,12 @@ import com.example.appgestiondeprojet.entity.User;
 import com.example.appgestiondeprojet.repository.ProjetRepository;
 import com.example.appgestiondeprojet.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProjetServiceImpl implements  IProjetservice {
@@ -39,8 +42,14 @@ public class ProjetServiceImpl implements  IProjetservice {
     }
 
     @Override
-    public void delete_projet(Long idprojet) {
-        projetrepo.deleteById(idprojet);
+    public ResponseEntity<String> delete_projet(Long idprojet) {
+        Optional<Projet> userOptional = projetrepo.findById(idprojet);
+        if (userOptional.isPresent()) {
+            projetrepo.deleteById(idprojet);
+            return ResponseEntity.ok("Projet supprimé avec succès");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Projet non trouvé");
+        }
     }
 
     @Override
